requirements.txt
requests>=2.31.0
pandas>=2.0.0
pydantic>=2.0.0
beautifulsoup4>=4.12.0
tqdm>=4.65.0
fastapi>=0.104.0
uvicorn>=0.24.0
tenacity>=8.2.0
python-dateutil>=2.8.0
rich>=13.0.0
lxml>=4.9.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
jinja2>=3.0.0
